+++ patternsleuth/src/resolvers/unreal/save_game.rs
@@ -12,7 +12,7 @@ use crate::resolvers::{ensure_one, impl_resolver_singleton};
     feature = "serde-resolvers",
     derive(serde::Serialize, serde::Deserialize)
 )]
-pub struct UGameplayStaticsSaveGameToMemory(pub usize);
+pub struct UGameplayStaticsSaveGameToMemory(pub u64);
 impl_resolver_singleton!(all, UGameplayStaticsSaveGameToMemory, |ctx| async {
     let patterns = [
         "48 89 5C 24 10 48 89 7C 24 18 55 48 8D AC 24 ?? FF FF FF 48 81 EC ?? 01 00 00 48 8B DA 48 8B F9 48 85 C9 0F 84 ?? 02 00 00 0F 57 C0 48 C7 85 ?? 00 00 00 00 00 00 00",
@@ -35,7 +35,7 @@ impl_resolver_singleton!(all, UGameplayStaticsSaveGameToMemory, |ctx| async {
     feature = "serde-resolvers",
     derive(serde::Serialize, serde::Deserialize)
 )]
-pub struct UGameplayStaticsSaveGameToSlot(pub usize);
+pub struct UGameplayStaticsSaveGameToSlot(pub u64);
 impl_resolver_singleton!(all, UGameplayStaticsSaveGameToSlot, |ctx| async {
     let patterns = [
         "48 89 5C 24 08 48 89 74 24 10 57 48 83 EC 40 ?? 8B ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? E8 ?? ?? FF FF 84 C0 74 58 E8 ?? ?? ?? ?? 48 8B ?? 48 8B ?? FF 52 ?? 4C 8B D0 48 85 C0 74 42 39 74 24 38 7E 3C 8B 53 08 ?? ?? ?? ?? ?? 0F 44 CE 85 C9 7E 2D",
@@ -58,7 +58,7 @@ impl_resolver_singleton!(all, UGameplayStaticsSaveGameToSlot, |ctx| async {
     feature = "serde-resolvers",
     derive(serde::Serialize, serde::Deserialize)
 )]
-pub struct UGameplayStaticsLoadGameFromMemory(pub usize);
+pub struct UGameplayStaticsLoadGameFromMemory(pub u64);
 impl_resolver_singleton!(all, UGameplayStaticsLoadGameFromMemory, |ctx| async {
     let patterns = [
         "48 89 5C 24 20 55 48 8D AC 24 10 FF FF FF 48 81 EC F0 01 00 00 83 79 08 00 48 8B D9 75 13 33 C0 48 8B 9C 24 18 02 00 00 48 81 C4 F0 01 00 00 5D C3 0F 57 C0 48 89 ?? 24 ?? 02 00 00 48 ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? 0F 11 45 ?? 0F 11 45 ?? 0F 11 45 ?? 0F 11 45 ?? 0F 11 45 ?? 0F 11 45 ?? 0F 11",
@@ -83,7 +83,7 @@ impl_resolver_singleton!(all, UGameplayStaticsLoadGameFromMemory, |ctx| async {
     feature = "serde-resolvers",
     derive(serde::Serialize, serde::Deserialize)
 )]
-pub struct UGameplayStaticsLoadGameFromSlot(pub usize);
+pub struct UGameplayStaticsLoadGameFromSlot(pub u64);
 impl_resolver_singleton!(all, UGameplayStaticsLoadGameFromSlot, |ctx| async {
     let patterns = [
         "48 8B C4 55 ?? 48 8D A8 ?? FE FF FF 48 81 EC ?? 02 00 00 48 89 ?? 08 33 ?? 48 89 ?? 10 48 8B ?? 4C 89 70 E8 44 8B F2 48 89 ?? 24 40 48 89 ?? 24 48 E8 ?? ?? ?? ?? 48 8B C8 4C 8B 00 41 FF 50 40 ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? ?? 48 8D 35",
@@ -109,7 +109,7 @@ impl_resolver_singleton!(all, UGameplayStaticsLoadGameFromSlot, |ctx| async {
     feature = "serde-resolvers",
     derive(serde::Serialize, serde::Deserialize)
 )]
-pub struct UGameplayStaticsDoesSaveGameExist(pub usize);
+pub struct UGameplayStaticsDoesSaveGameExist(pub u64);
 impl_resolver_singleton!(all, UGameplayStaticsDoesSaveGameExist, |ctx| async {
     let patterns = [
         "48 89 5C 24 08 57 48 83 EC 20 8B FA 48 8B D9 E8 ?? ?? ?? ?? 48 8B C8 4C 8B 00 41 FF 50 ?? 48 85 C0 74 3D 83 7B 08 00 4C 8B 00 4D 8B 48 ?? 74 16 48 8B 13 44 8B C7 48 8B C8 48 8B 5C 24 30 48 83 C4 20 5F 49 FF E1 48 8D 15 ?? ?? ?? ?? 44 8B C7 48 8B C8 48 8B 5C 24 30 48 83 C4 20 5F 49 FF E1 48 8B 5C 24",
